+++ src/chess.zig
@@ -2,6 +2,7 @@ const color = @import("chess/color.zig");
 const square = @import("chess/square.zig");
 
 pub const BitBoard = @import("chess/BitBoard.zig");
+pub const BitPieceMap = @import("chess/BitPieceMap.zig");
 pub const Color = color.Color;
 pub const File = square.File;
 pub const Piece = @import("chess/piece.zig").Piece;