+++ CHANGELOG.md
@@ -7,7 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+#### Chess module
+
+- `BitPieceMap` board representation.
+  - `Extended` version for more info.
+
 ## [0.3.0] - 2024-08-13
+
+- `BitBoard` `isEmpty` method.
 
 ### Chess module